Refactor continuous aggregate code
