+++ tsl/src/continuous_aggs/create.h
@@ -11,14 +11,10 @@
 #include "with_clause_parser.h"
 #include "ts_catalog/continuous_agg.h"
 
-#define CONTINUOUS_AGG_CHUNK_ID_COL_NAME "chunk_id"
-
 DDLResult tsl_process_continuous_agg_viewstmt(Node *node, const char *query_string, void *pstmt,
 											  WithClauseResult *with_clause_options);
 
 extern void cagg_flip_realtime_view_definition(ContinuousAgg *agg, Hypertable *mat_ht);
 extern void cagg_rename_view_columns(ContinuousAgg *agg);
-
-extern Datum tsl_cagg_try_repair(PG_FUNCTION_ARGS);
 
 #endif /* TIMESCALEDB_TSL_CONTINUOUS_AGGS_CAGG_CREATE_H */
